% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_alroy}
\alias{C_alroy}
\title{Alroy's (2015) modification of the Forbes index}
\usage{
C_alroy(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist object with the pairwise association values.
}
\description{
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
\references{
Alroy J. (2015) A new twist on a very old binary similarity coefficient.
Ecology 96: 575-586.
}
