% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_combo}
\alias{C_combo}
\title{Number of unique species combinations}
\usage{
C_combo(m, clean = TRUE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{clean}{Logical. Do you want to remove zero-sum rows or columns? When the index needs
the "d" fraction (sites where no species occurs), only zero-sum rows are removed.}
}
\value{
A single number, the number of unique species combinations.
}
\description{
This function follows the code from EcoSimR package by Gotelli, Hard and Ellison,
specifically, it re-uses their 'species_combo' function.
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
