% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_dale}
\alias{C_dale}
\title{Association metric of Dale (1999)}
\usage{
C_dale(m, lst = FALSE, fun = FALSE)
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
This is a metric described on page 147 of Dale (1999).
}
\references{
Dale M.T.D (1999) Spatial pattern analysis in plant ecology.
Cambridge University Press.
}
