% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_dale}
\alias{C_dale}
\title{Association metric of Dale (1999)}
\usage{
C_dale(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
This is a metric described on page 147 of Dale (1999).
}
\references{
Dale M.T.D (1999) Spatial pattern analysis in plant ecology.
Cambridge University Press.
}
