% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_jacc}
\alias{C_jacc}
\title{Pairwise Jaccard associations among species in a community matrix}
\usage{
C_jacc(m, lst = FALSE, fun = FALSE)
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
Pairwise Jaccard associations among species in a community matrix
}
