% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_Pearson}
\alias{C_Pearson}
\title{Pairwise pearson correlations among species in community matrix}
\usage{
C_Pearson(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can 
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied 
to all pairwise values.}
}
\value{
A list or data.frame objects with the pairwise association values.
}
\description{
Pairwise pearson correlations among species in community matrix
}
