% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_ratio}
\alias{C_ratio}
\title{Variance ratio of Schluter (1984)}
\usage{
C_ratio(m, clean = TRUE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{clean}{Logical. Do you want to remove zero-sum rows or columns? When the index needs
the "d" fraction (sites where no species occurs), only zero-sum rows are removed.}
}
\value{
A single number, the variance ratio.
}
\description{
This is the classical variance ratio function; the function is applicable to
both incidence and abundance matrices.
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
\references{
Schluter, D. 1984. A variance test for detecting species associations,
with some example applications. Ecology 65: 998-1005.
}
