% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_seg}
\alias{C_seg}
\title{The classical 'C-score' (C_seg) seggregation metric of Stone & Roberts (1990),
and its scaled version}
\usage{
C_seg(m, scale = TRUE, clean = TRUE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{scale}{Should the raw metric be scaled using the total number of possible
site combinations?}

\item{clean}{Logical. Do you want to remove zero-sum rows or columns? When the index needs
the "d" fraction (sites where no species occurs), only zero-sum rows are removed.}
}
\value{
A dist object with the pairwise association values.
}
\description{
The scaled version divides the raw C-score by the number of possible site
combinations, which is `n(n-1)/2` according to Ulrich & Gotelli (2013). Also
note that their scaling formula also includes `m`, the number of species. Thus,
Ulrich & Gotelli's (2013) formula for the scaled C-score is actually
`4*(C_seg)/(m*n*(m-1)(n-1))`. For two species (`m=2`) this reduces to  `C_seg / ((n*(n-1))/2)`,
which is the formula that I use here for the scaled version.
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
\references{
Stone L. & Roberts A. (1990) The checkerboard score and
species distributions. Oecologia 85: 74-79.

McNickle G.G. et al. (2018) Checkerboard score-area relationships
reveal spatial scales of plant community structure. Oikos 127: 415-426.

Ulrich W. & Gotelli N.J. (2013) Pattern detection in null model analysis.
Oikos 122: 2-18.
}
