% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_sim}
\alias{C_sim}
\title{Pairwise Simpson seggregation among species in a community matrix}
\usage{
C_sim(m, lst = FALSE, fun = FALSE)
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
Pairwise Simpson seggregation among species in a community matrix
}
