% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_sor}
\alias{C_sor}
\title{Pairwise Sorensen associations among species in a community matrix}
\usage{
C_sor(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
Arita (2017) also quotes this to be the "index of co-incidence" of Dice (1945).
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
\references{
Arita H. (2015) Multisite and multispecies measures of overlap,
co-occurrence and co-diversity. Ecography 40: 709-718.

Dice L.R. (1945) Measures of the amount of ecological association
between species. Ecology 94: 2403-2414.
}
