% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_Stone}
\alias{C_Stone}
\title{Classical 'C-score' metric of Stone -- seggregation among species in a community matrix}
\usage{
C_Stone(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can 
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied 
to all pairwise values.}
}
\value{
A list or data.frame objects with the pairwise association values.
}
\description{
Classical 'C-score' metric of Stone -- seggregation among species in a community matrix
}
\references{
Stone L. & Roberts A. (1990) The checkerboard score and 
species distributions. Oecologia 85: 74-79.
}
