% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_tog}
\alias{C_tog}
\title{'Togetherness' metric of Stone & Roberts (1992), and its scaled version}
\usage{
C_tog(m, lst = FALSE, fun = FALSE, scale = FALSE)
}
\arguments{
\item{scale}{Should the raw metric be scaled using the total number of possible
site combinations?}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
'Togetherness' metric of Stone & Roberts (1992), and its scaled version
}
\references{
Stone L. & Roberts A. (1992) Competitive exclusion, or species
aggregation? An aid in deciding. Oecologia 91: 419-424.

Ulrich W. & Gogelli N.J. (2013) Pattern detection in null model
analysis. Oikos 122: 2-18.
}
