% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_tog}
\alias{C_tog}
\title{'Togetherness' (C_tog) metric of Stone & Roberts (1992), and its scaled version}
\usage{
C_tog(m, scale = TRUE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{scale}{Should the raw metric be scaled using the total number of possible
site combinations?}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
The scaled version divides the raw togetherness by the number of possible site
combinations, which is `n(n-1)/2` according to Ulrich & Gotelli (2013). Also
note that their scaling formula  includes `m`, the number of species. Thus,
Ulrich & Gotelli's (2013) formula for the scaled C-score is actually
`4*(C_tog)/(m*n*(m-1)(n-1))`. For two species (`m=2`) this reduces to  `C_tog / ((n*(n-1))/2)`,
which is the formula that I use here for the scaled version.
Note: Before the calculation, species (rows) with zero occurrences are removed from m.
}
\references{
Stone L. & Roberts A. (1992) Competitive exclusion, or species
aggregation? An aid in deciding. Oecologia 91: 419-424.

Ulrich W. & Gogelli N.J. (2013) Pattern detection in null model
analysis. Oikos 122: 2-18.
}
