% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_Ulrich}
\alias{C_Ulrich}
\title{'C-score' metric of Stone & Roberts, scaled according to Ulrich & Gotelli 2012}
\usage{
C_Ulrich(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can 
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied 
to all pairwise values.}
}
\value{
A list or data.frame objects with the pairwise association values.
}
\description{
'C-score' metric of Stone & Roberts, scaled according to Ulrich & Gotelli 2012
}
\references{
Stone L. & Roberts A. (1990) The checkerboard score and 
species distributions. Oecologia 85: 74-79.

McNickle G.G. et al. (2018) Checkerboard score-area relationships
reveal spatial scales of plant community structure. Oikos 127: 415-426.
}
