% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_abundance_matrices.R
\name{CA_bray}
\alias{CA_bray}
\title{Bray-Curtis dissimilarity matrix}
\usage{
CA_bray(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist object with the pairwise values.
}
\description{
Calculates all Bray-Curtis dissimilarities for all pairs of species in a community matrix.
This is a wrapper to a specific call to the 'vegdist' function in 'vegan'.
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
\references{
Legendre P. & De Caceres M. (2013) Beta diversity as the variance of community
data: dissimilarity coefficients and partitioning. Ecology Letters 16: 951-963.
}
