% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_abundance_matrices.R
\name{CA_chi}
\alias{CA_chi}
\title{Chi-square distance matrix}
\usage{
CA_chi(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist object with the pairwise values.
}
\description{
Calculates chi-square distances for all pairs of species in a community matrix.
In particular,
the abundance data are chi-square-transformed, and then used for calcluation
of Euclidean distances.
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
\references{
Legendre P. & De Caceres M. (2013) Beta diversity as the variance of community
data: dissimilarity coefficients and partitioning. Ecology Letters 16: 951-963.

Legendre P. & Legendre L. (2012) Numerical Ecology. Elsevier. pp 265-335.
}
