% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_abundance_matrices.R
\name{CA_ruz}
\alias{CA_ruz}
\title{Abundance-based Jaccard (a.k.a. Ruzicka) dissimilarity matrix}
\usage{
CA_ruz(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist or data.frame objects with the pairwise values.
}
\description{
Calculates abundance-based Jaccard dissimilarity matrix for a community matrix.
This is a wrapper to a specific call to the 'vegdist' function in 'vegan'.
}
\references{
Legendre P. & De Caceres M. (2013) Beta diversity as the variance of community
data: dissimilarity coefficients and partitioning. Ecology Letters 16: 951-963.
}
