% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_matrix_measures.R
\name{CA_seg}
\alias{CA_seg}
\title{Abundance-based checkerboard score}
\usage{
CA_seg(m, lst = FALSE, fun = FALSE, scale = TRUE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}

\item{scale}{Should the raw metric be scaled using the total number of possible
site combinations?}
}
\description{
This metric is described in Ulrich & Gotelli (2010).
}
\references{
Ulrich & Gotelli (2010) Null model analysis of species associations
using abundance data. 91: 3384-3397.
}
