% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_ecosim_functions.R
\name{cooc_null_model_spasm}
\alias{cooc_null_model_spasm}
\title{Co-Occurrence Null model - modified verstion from EcoSimR}
\usage{
cooc_null_model_spasm(speciesData, algo = "sim9", metric, nReps = 1000,
  saveSeed = FALSE, burn_in = 500, algoOpts = list(),
  metricOpts = list(), suppressProg = FALSE, summaryF)
}
\arguments{
\item{speciesData}{a dataframe in which rows are species, columns are sites,
and the entries indicate the absence (0) or presence (1) of a species in a
site. Empty rows and empty columns should not be included in the matrix.}

\item{algo}{the algorithm to use, must be "sim1", "sim2", "sim3", "sim4", "sim5", "sim6", "sim7", "sim8", "sim9", "sim10"; default is "sim9".}

\item{metric}{the metric used to calculate the null model: the choice is unrestricted}

\item{nReps}{the number of replicate null assemblages to create; default is 1000 replicates.}

\item{saveSeed}{TRUE or FALSE. If TRUE the current seed is saved so the simulation can be repeated; default is FALSE.}

\item{burn_in}{The number of burn_in iterations to use with the simFast algorithm; default is 500 burn-in replicates.}

\item{algoOpts}{a list containing all the options for the specific algorithm you want to use.  Must match the algorithm given in the `algo` argument.}

\item{metricOpts}{a list containing all the options for the specific metric you want to use.  Must match the metric given in the `metric` argument.}

\item{suppressProg}{TRUE or FALSE. If true, display of the progress bar in the console is suppressed; default is FALSE. This setting is useful for creating markdown documents with `knitr`.}
}
\description{
Create a Co-Occurrence null model. This is a modified version of the function
"cooc_null_model"
from EcoSimR, with no restrictions on the metric that is applied to the community matrix.
}
