% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{create_fleet}
\usage{
create_fleet(eq_f = NA, length_50_sel = NA, length_95_sel = NA,
  mpa_reaction = "concentrate", fish)
}
\arguments{
\item{fish}{}
}
\value{
a fleet object
}
\description{
create_fleet
}
\examples{
create_fleet(eq_f = 2,length_50_sel = 25, length_95_sel = 27, fish = bluefish)

}
