% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manager.R
\name{create_manager}
\alias{create_manager}
\title{create manager}
\usage{
create_manager(year_mpa = 15)
}
\arguments{
\item{year_mpa}{}
}
\value{
a manager object
}
\description{
create manager
}
\examples{
create_manager(year_mpa = 12)
}
