% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.Atmar}
\alias{data.Atmar}
\title{Atmar & Patterson (1995) incidence matrices}
\format{
An object of class \code{list} of length 290.
}
\source{
Atmar W. & Patterson B. D. (1995) The nestedness temperature calculator:
a Visual Basic  program,  including  294  presence absence matrices. AICS Research,
Univ. Park, NM and Field Museum, Chicago, http://aics-research.com/nestedness/.
}
\usage{
data.Atmar
}
\description{
This is a list of site by species binary *incidence* (0 or 1) matrices.
Note that after all cleaning procedures, the final number of matrices is 290 (not 294).
Each matrix in the list has a name that enables it to be linked back to the
original dataset in the publication.
}
\keyword{datasets}
