% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.Harvard}
\alias{data.Harvard}
\title{Orwig, Foster & Ellison (2015) Harvard forest plot}
\format{
An object of class \code{data.frame} with 77536 rows and 4 columns.
}
\source{
Owig D., Foster D. & Ellison A. (2015) Harvard Forest CTFS-ForestGEO
Mapped Forest Plot since 2014. Harvard Forest Data Archive: HF253.
}
\usage{
data.Harvard
}
\description{
This is a set of all 'alive' trees from the file hf253-03-trees-2014.csv, downloaded from
http://harvardforest.fas.harvard.edu:8080/exist/apps/datasets/showData.html?id=hf253
This represents the 2014 census.
}
\keyword{datasets}
