% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.journals}
\alias{data.journals}
\title{Data on 3879 papers published in American Nautralist, Ecology, and Ecography
in years 1995, 1999, 2003, 2007, 2011, and 2019}
\format{
An object of class \code{data.frame} with 3856 rows and 26 columns.
}
\source{
http://apps.webofknowledge.com/
}
\usage{
data.journals
}
\description{
I created the data by
reading 3879 abstracts and manually classifying them into thematic categories.
The abstracts were taken from Clarivate Web of Science. Metadata are in the "data" folder.
}
\keyword{datasets}
