% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.Ulrich}
\alias{data.Ulrich}
\title{Ulrich & Gotelli (2010) abundance matrices}
\format{An object of class \code{list} of length 186.}
\source{
Ulrich W. & Gotelli N.J. (2010) Null model analysis of species associations
using abundance data. Ecology 91: 3384-3397.
}
\usage{
data.Ulrich
}
\description{
This is a list of 186 site by species *abundance* matrices.
Each matrix in the list has a name that enables it to be linked back to the
original dataset in the publication.
}
\keyword{datasets}
