% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.WOS}
\alias{data.WOS}
\title{Number of papers on biodiversity + a given category on Clarivate Web of Science}
\format{
An object of class \code{data.frame} with 883 rows and 6 columns.
}
\source{
http://apps.webofknowledge.com/
}
\usage{
data.WOS
}
\description{
The search was done on 19 Sep 2019. The search categories were the given term (in the data)
AND "biodiversity".
}
\keyword{datasets}
