% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_communities.R
\name{dpoint.MVN.image}
\alias{dpoint.MVN.image}
\title{Funciton that generates the MVN density as a pixel image}
\usage{
dpoint.MVN.image(var, x.centr, y.centr)
}
\arguments{
\item{var}{Variance of the PDF, which is the same in x and y direction. Note: covariance is 0.}

\item{x.centr}{x coordinate of the mean in the 2D space}

\item{y.centr}{y coordinate of the mean in the 2D space}
}
\description{
Funciton that generates the MVN density as a pixel image
}
