% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_communities.R
\name{dpoint.MVN}
\alias{dpoint.MVN}
\title{Bivariate probablity density function in a 2D square}
\usage{
dpoint.MVN(x, y, var, x.centr, y.centr)
}
\arguments{
\item{x}{x Coordinates of points whose PD should be evaluated.}

\item{y}{y Coordinates of points whose PD should be evaluated.}

\item{var}{Variance of the PDF, which is the same in x and y direction. Note: covariance is 0.}

\item{x.centr}{x coordinate of the mean in the 2D space}

\item{y.centr}{y coordinate of the mean in the 2D space}
}
\description{
Bivariate probablity density function in a 2D square
}
