% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-costs.R
\name{estimate_costs}
\alias{estimate_costs}
\title{Title}
\usage{
estimate_costs(cost, fish, fleet, msy, e_msy, b_msy, b_v_bmsy_oa,
  p_response, sim_years = 100, burn_year = 25, num_patches = 1,
  use = "fit", lags = 0)
}
\arguments{
\item{use}{}
}
\value{
an estimate of costs
}
\description{
Title
}
