% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSplines.R
\name{fitted.sparseSpline}
\alias{fitted.sparseSpline}
\title{Extract fitted values from a \code{sparseSpline} object}
\usage{
\method{fitted}{sparseSpline}(object, which.lam = 10, ...)
}
\arguments{
\item{object}{an object of class \code{sparseSpline}, as produced by \code{\link{spam}}}

\item{which.lam}{an optional value for the index of the tuning parameter sequence to produce. If \code{NULL}, fitted values for all values of the tuning parameter}

\item{...}{optional additional arguments for method compatability}
}
\value{
A matrix of dimension Nxd or an array of dimension N x nlambda x d if \code{which.lam=NULL}.
}
\description{
Extract fitted values from a \code{sparseSpline} object
}
\examples{
someData = generateSPASM(100,3,4)
SPAMfit = spam(someData$y, someData$x, returnall=TRUE)
SPAMpreds = fitted(SPAMfit)
}
