% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateSPASM}
\alias{generateSPASM}
\title{Generate data from an additive state space model}
\usage{
generateSPASM(N, p, d, sig.x = 0.1, sig.y = 0.1)
}
\arguments{
\item{N}{the number of observed time points}

\item{p}{the dimension of the hidden state vector}

\item{d}{the dimension of the observation vector}

\item{sig.x}{scalar which multiplies the hidden standard normal noise}

\item{sig.y}{scalar which multiplies the standard normal noise added to the observations}
}
\value{
a list with components
\describe{
\item{y}{the observations, nxd}
\item{x}{the hidden states, nxp}
\item{Tt}{the state transition matrix, pxp}
\item{HHt}{the observation noise variance, dxd}
\item{GGt}{the hidden noise variance, pxp}
}
}
\description{
This function generates data and returns components so as to easily test the Laplace-Gaussian filter and the sparse spline additive models. Data is generated as
\deqn{y_{t} = Z sin(z * x_{t} / 2\pi) + \sigma_{y} * N(0, I)}
\deqn{x_{t+1} = Tt x_t + \sigma_x * N(0,I)}
where \eqn{Z} is a dxp matrix of standard uniform random variables and \eqn{z=[1 ... p]}.
}
\examples{
generateSPASM(100,3,4)
}

