% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubalek.R
\name{hubalek}
\alias{hubalek}
\title{Hubalek's indices calculated all at once}
\usage{
hubalek(x1, x2)
}
\arguments{
\item{x1}{A binary vector.}

\item{x2}{A binary vector of the same legnth as x1.}
}
\value{
A dist or data.frame objects with the pairwise association values. Letters and
numbers correspond to those in Hubalek's paper.
}
\description{
This is a function that calculates all of the indices listed in Hubalek (1982).
}
\references{
Hubalek Z. (1982) Coefficients of association and similarity,
based on binary (presence-absence) data: and evaluation. Biol. Rev., 57: 669-689.
}
