% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgf.R
\name{lgf1}
\alias{lgf1}
\title{First order LGF}
\usage{
lgf1(Y, out.spam, Tt, HHinv, GGt, x0, P0)
}
\arguments{
\item{Y}{the observed data, a Nxd matrix}

\item{out.spam}{the sparse additive spline model associated with the observation equation, as produced by \code{\link{spam}}}

\item{Tt}{the state transition matrix, pxp}

\item{HHinv}{the inverse of the observation noise variance matrix}

\item{GGt}{the transition noise variance matris}

\item{x0}{prior mean of the state variables}

\item{P0}{prior variance of the state variables}
}
\value{
A list with components \code{xt} and \code{xtt}, the predicted and filtered state means respectively and \code{Pt} and \code{Ptt}, the associated variances.
}
\description{
Given a sparse-spline representation of the observation equation, this function performs filtering to derive the conditional distributions of the filtered (\eqn{x_{t|t}}) and predicted (\eqn{x_{t|t-1}}) hidden states.
}
\examples{
someData = generateSPASM(100,3,4)
SPAMfit = spam(someData$y, someData$x)
filt = lgf1(someData$y, SPAMfit, someData$Tt,
  solve(someData$HHt), someData$GGt, rep(0,3), diag(1,3))
matplot(someData$x, ty='l', col=1, lty=1)
matlines(filt$xtt, col=2, lty=3)
}

