% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_adults.R
\name{move_adults}
\alias{move_adults}
\title{\code{move_adults} moves adult fish via a movement function}
\usage{
move_adults(here_pop)
}
\arguments{
\item{here_pop}{}
}
\value{
relocated adult fish
}
\description{
\code{move_adults} moves adult fish via a movement function
}
\examples{
move_adults(pop)
}
