% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_counterfactual.R
\name{mpa_counterfactual}
\alias{mpa_counterfactual}
\title{\code{mpa_counterfactual} compares a fishery with and without an MPA}
\usage{
mpa_counterfactual(fish, fleet, year_mpa, mpa_size, sim_years, num_patches,
  burn_years)
}
\arguments{
\item{burn_years}{}
}
\value{
an object comparing runs
}
\description{
\code{mpa_counterfactual} compares a fishery with and without an MPA
}
