% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{N_checker}
\alias{N_checker}
\title{Number of checkerboard species pairs}
\usage{
N_checker(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A single number, the number of checkerboard species pairs.
}
\description{
This function follows the code from EcoSimR package by Gotelli, Hard and Ellison,
specifically, it re-uses their 'checker' function.
}
