% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{N_combo}
\alias{N_combo}
\title{Number of unique species combinations}
\usage{
N_combo(m)
}
\value{
A single number, the number of unique species combinations.
}
\description{
This function follows the code from EcoSimR package by Gotelli, Hard and Ellison,
specifically, it re-uses their 'species_combo' function.
}
