% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.R
\name{N_connect}
\alias{N_connect}
\title{Network connectance}
\usage{
N_connect(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A single number, network connectance.
}
\description{
Network connectance, defined as the proportion of all possible links in a network, adopted
from Carsten Dormann's bipartite package.
}
\references{
Dormann et al. (2009) Indices, graphs and null models: analyzing
bipartite ecological networks. The Open Ecology Journal, 2: 7-24.
}
