% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgf.R
\name{negHessL}
\alias{negHessL}
\title{Calculates the negative Hessian of \eqn{l(x)}}
\usage{
negHessL(x, prev.state, y, out.spam, HHinv)
}
\arguments{
\item{x}{p-vector of current states (to minimize over)}

\item{prev.state}{list containing mean and variance of \eqn{\hat{p}(x_{t-1} | y_{t-1})} (as produced by \code{\link{posteriorPredictive}}}

\item{y}{d-vector of observations at time t}

\item{out.spam}{list of fitted splines for the observation equation, as produced by \code{\link{spam}}}

\item{HHinv}{inverse of the observation error variance}
}
\value{
The negative Hessian of the (approximate) \eqn{l'(x)}, (pxp matrix). Note that this is the negative of the Hessian of \eqn{l(x)}, not \eqn{-l(x)}.
}
\description{
Calculates the negative Hessian of \eqn{l(x)}
}
\details{
This function calculates
\deqn{-l''(x) = - d^2/dx^2 log p(y_{t} | x_{t}) p(x_{t} | y_{1:t-1}).}
As the second distribution in the above product is approximated with a Gaussian, the result is also an approximation.
}
