% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgf.R
\name{negL}
\alias{negL}
\title{Calculates negative \eqn{l(x_t)} up to an additive constant}
\usage{
negL(x, prev.state, y, out.spam, HHinv)
}
\arguments{
\item{x}{p-vector of current states (to minimize over)}

\item{prev.state}{list containing mean and variance of \eqn{\hat{p}(x_{t-1} | y_{t-1})} (as produced by \code{\link{posteriorPredictive}}}

\item{y}{d-vector of observations at time t}

\item{out.spam}{list of fitted splines for the observation equation, as produced by \code{\link{spam}}}

\item{HHinv}{inverse of the observation error variance}
}
\value{
The negative of the (approximate) \eqn{l(x)}
}
\description{
Calculates negative \eqn{l(x_t)} up to an additive constant
}
\details{
This function calculates
\deqn{-l(x) = -log p(y_{t} | x_{t}) p(x_{t} | y_{1:t-1}).}
As the second distribution in the above product is approximated with a Gaussian, the result is also an approximation. Furthermore, as the first order LGF (\code{\link{lgf1}}) requires minimization over \eqn{-l(x)}, this function calculates it only to a constant of porportionality (independent of x)).
}

