% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_simulations.R
\name{null_model_abund_spasm}
\alias{null_model_abund_spasm}
\title{Abundance Null model}
\usage{
null_model_abund_spasm(m, metric, nReps = 1000, summaryF)
}
\description{
Uses the IT algorithm by Ulrich & Gotelli (2010)
}
