% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_simulations.R
\name{null_model_cooc_spasm}
\alias{null_model_cooc_spasm}
\title{Cooccurrence (binary presence-absence) null model - modified verstion from EcoSimR}
\usage{
null_model_cooc_spasm(speciesData, algo = "sim9", metric, nReps = 1000,
  burn_in = 500, suppressProg = FALSE, summaryF, ...)
}
\arguments{
\item{speciesData}{a dataframe in which rows are species, columns are sites,
and the entries indicate the absence (0) or presence (1) of a species in a
site. Empty rows and empty columns should not be included in the matrix.}

\item{algo}{the algorithm to use, must be "sim1", "sim2", "sim3", "sim4", "sim5", "sim6", "sim7", "sim8", "sim9", "sim10"; default is "sim9".}

\item{metric}{the metric used to calculate the null model: the choice is unrestricted}

\item{nReps}{the number of replicate null assemblages to create; default is 1000 replicates.}

\item{burn_in}{The number of burn_in iterations to use with the simFast algorithm; default is 500 burn-in replicates.}

\item{suppressProg}{TRUE or FALSE. If true, display of the progress bar in the console is suppressed; default is FALSE. This setting is useful for creating markdown documents with `knitr`.}
}
\description{
Create a Co-Occurrence null model. This is based on the function
"cooc_null_model" from EcoSimR, with has been modified to have
no restrictions on the metric that is applied to the community matrix.
}
