% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pattern_measures.R
\name{pcf.r}
\alias{pcf.r}
\title{Bivariate pair correlation function for a given distance r}
\usage{
\method{pcf}{r}(X, spec.1, spec.2, n.iter, radius.seq)
}
\arguments{
\item{X}{Object of class 'ppp' (from 'spatstat' packge). This is a marked 
point pattern in which each mark gives identity of a species. Thus, the object
must have at least 2 species.}

\item{spec.1}{Character string indicating name of species 1 in the 'ppp' object.}

\item{spec.2}{Character string indicating name of species 2 in the 'ppp' object.}

\item{n.iter}{Number of iterations for simulation from a homogeneous Poisson
point process.}

\item{radius.seq}{Numeric vector. Sequence of radii for which the pcf function 
should be calculated}
}
\value{
A list.
}
\description{
Bivariate pair correlation function for a given distance r
}
\references{
Wiegand T. et al. (2007) Species associations in a heterogeneous
Sri Lankan dipterocarp forest. American Naturalist 170: E77-E95.

Wang X. et al. (2010) Species associations in an old-growth temperate
forest in north-eastern China. Journal of Ecology, 95: 674-686.

Wiegand T. & Moloney K. A. (2014) Handbook of spatial point pattern
analysis in ecology. CRC press.
}
