% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_patterns_simulated_communities.R
\name{PDFtexp}
\alias{PDFtexp}
\title{Truncated exponential probability density function}
\usage{
PDFtexp(d, alpha, dlim = c(0, sqrt(2)))
}
\description{
Truncated exponential probability density function
}
