% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pattern_measures.r
\name{plot.pp.pcf}
\alias{plot.pp.pcf}
\title{Bivariate nearest neighbor distance function (D) and K-function (K)}
\usage{
plot.pp.pcf(x)
}
\arguments{
\item{PM}{A list object returned by the PM.multispec function.}
}
\value{
A ggplot figure.
}
\description{
Bivariate nearest neighbor distance function (D) and K-function (K)
}
