% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_point_patterns.R
\name{PM.multispec}
\alias{PM.multispec}
\title{Multi-species P-M classfication of Wiegand et al. (2007)}
\usage{
PM.multispec(X, n.iter, radius)
}
\arguments{
\item{X}{Object of class 'ppp' (from 'spatstat' packge). This is a marked
point pattern in which each mark gives identity of a species. Thus, the object
must have at least 2 species.}

\item{n.iter}{Number of iterations for simulation from a homogeneous Poisson
point process.}

\item{r}{Radius for which the P and M values should be calculated.}
}
\value{
A list.
}
\description{
Still under development - use it cautiously
}
\references{
Wiegand T. et al. (2007) Species associations in a heterogeneous
Sri Lankan dipterocarp forest. American Naturalist 170: E77-E95.

Wang X. et al. (2010) Species associations in an old-growth temperate
forest in north-eastern China. Journal of Ecology, 95: 674-686.

Wiegand T. & Moloney K. A. (2014) Handbook of spatial point pattern
analysis in ecology. CRC press.
}
