% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgf.R
\name{posteriorPredictive}
\alias{posteriorPredictive}
\title{Calculates the convolution of two Gaussian distributions}
\usage{
posteriorPredictive(x, P, Tt, GGt)
}
\arguments{
\item{x}{the mean of \eqn{x_{t|t}}}

\item{P}{the variance of \eqn{x_{t|t}}}

\item{Tt}{the state transition matrix}

\item{GGt}{the variance of the transition noise}
}
\value{
a list with components \code{x}, the mean of \eqn{x_{t+1|t}}, \code{P} the variance of \eqn{x_{t+1|t}} and \code{Pinv}, the inverse of \code{P}
}
\description{
Calculates the convolution of two Gaussian distributions
}
\details{
This function calculates
\deqn{p(x_{t+1} | y_{t}) = \int dx_{t} p(x_{t+1}| x_{t}) p(x_{t} | y_{1:t})}
assuming that both distributions on the right are Gaussian. This function also assumes that the mean of \eqn{x_{t+1|t}} is linear in \eqn{x_{t}}:
\deqn{p(x_{t+1} | x_{t}) = N(Tt x_{t|t}, GGt).}
The result is therefore:
\deqn{p(x_{t+1} | y_{t}) = N(Tt x_{t|t}, Tt P Tt' + GGt).}
}
\examples{
posteriorPredictive(rnorm(3), diag(1,3), matrix(rnorm(9),3), diag(1,3))
}

