% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ppp.to.comm}
\alias{ppp.to.comm}
\title{Conversion of point pattern to community matrix Y}
\usage{
ppp.to.comm(comm.ppp, dim.yx, plotstack = FALSE, clean = FALSE)
}
\arguments{
\item{comm.ppp}{A marked point pattern of spatstat's 'ppp', where marks identify each
point (individual) to a species.}

\item{dim.yx}{Number of cells along the Y and X dimension of the raster.}

\item{plotstack}{Logical. Should the raster for each species be plotted?}

\item{clean}{Logical. Should only sites with positive abundance of at least one
species be returned?}
}
\value{
A list containing the raster stack with rasters for each species,
the abundance matrix Y, spatial coordinates of each site (grid cell), and
number of cells.
}
\description{
Conversion of point pattern to community matrix Y
}
