% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_communities.R
\name{rpoint.MVN}
\alias{rpoint.MVN}
\title{Function that generates n points in a 2D space with a bivariate normal PDF,
given the variance, and xy coordinates of the centre of the PDF.}
\usage{
rpoint.MVN(n, var, x.centr, y.centr)
}
\arguments{
\item{n}{number of points}

\item{var}{Variance of the PDF, which is the same in x and y direction. Note: covariance is 0.}

\item{x.centr}{x coordinate of the mean in the 2D space}

\item{y.centr}{y coordinate of the mean in the 2D space}
}
\description{
Function that generates n points in a 2D space with a bivariate normal PDF,
given the variance, and xy coordinates of the centre of the PDF.
}
