% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_communities.R
\name{sim.comm.centered}
\alias{sim.comm.centered}
\title{Function that simulates a community of several species.
The simulation proceeds in 2 steps:
1. Genreate set of 'mother points' from the bivariate normal, with a given var.intersp.
2. Treat each 'mother point' as a centre of a cluster of points of a given species,
   with var.consp as the "spread" of the cluster.}
\usage{
sim.comm.centered(abund.vect, var.consp, var.intersp)
}
\arguments{
\item{abund.vect}{vector of (integer) abundances of species in a community}

\item{var.consp}{conspecific variance, i.e. spread of points in a single species}

\item{var.intersp}{interspecific variance, i.e. spread of the mother points}
}
\description{
Function that simulates a community of several species.
The simulation proceeds in 2 steps:
1. Genreate set of 'mother points' from the bivariate normal, with a given var.intersp.
2. Treat each 'mother point' as a centre of a cluster of points of a given species,
   with var.consp as the "spread" of the cluster.
}
