% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_patterns_simulated_communities.R
\name{sim.comm.jittered}
\alias{sim.comm.jittered}
\title{Function that simulates a "jittered" community}
\usage{
sim.comm.jittered(abund.vect, var.consp, var.intersp)
}
\arguments{
\item{abund.vect}{vector of (integer) abundances of species in a community}

\item{var.consp}{conspecific variance, i.e. spread of points in the master species}

\item{var.intersp}{interspecific variance, i.e. width of the density kernel}
}
\description{
The simulation proceeds as:
1. Generate point pattern for the most abundant (master) species,
   with a given variance 'var.consp' giving the spread of the points
   along the x and y axis
2. Generate smoothed density of the master points, with a
   given gaussian kernel width 'var.intersp'. The steeper the kernel,
   the closer will be the points of the other species to the mother species.
3. For each other species in the vector, draw points from the smoothed
   density surface, with a given abundance
Note: Master species is the most abundant species, i.e. species 1.
}
