% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_patterns_simulated_communities.R
\name{sim.pair}
\alias{sim.pair}
\title{Two point patterns with a given attraction or repulsion}
\usage{
sim.pair(abund.vect, var.consp, alpha, plot.comm = FALSE)
}
\arguments{
\item{abund.vect}{2-element vector with abundances (# of individuals) of the
two species for which the pattern will be simulated.}

\item{alpha}{inter-specific association (attraction) or repulsion. Values < 0 are
for repulsion, 0 is no relationship, values > 0 are attraction}

\item{plot.comm}{should the pattern be plotted?}

\item{var.intersp}{interspecific variance, i.e. width of the density kernel.
It's a positive number, and the closer is the value to 0 the stronger the
intraspecific aggregation.}
}
\description{
Two point patterns with a given attraction or repulsion
}
