% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_simulations.R
\name{sim9_spasm}
\alias{sim9_spasm}
\title{Modified sim9 algorithm to handle spasm's functions, taken from EcoSimR}
\usage{
sim9_spasm(speciesData, metric, nReps = 1000, burn_in = 0,
  suppressProg = TRUE, summaryF)
}
\description{
Modified sim9 algorithm to handle spasm's functions, taken from EcoSimR
}
