% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fishery.R
\name{sim_fishery}
\alias{sim_fishery}
\title{\code{sim_fishery} simulates an age structured spatially explicit
model forward with fleets etc.}
\usage{
sim_fishery(fish, fleet, manager, num_patches = 10, sim_years = 25,
  burn_year = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
a pop object with population and catch trajectories
}
\description{
\code{sim_fishery} simulates an age structured spatially explicit
model forward with fleets etc.
}
\examples{
sim_fishery(fish = fish, fleet = fleet,...)

}
