% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSplines.R
\name{spam}
\alias{spam}
\title{Estimates sparse (mulitivariate) additive model}
\usage{
spam(Y, X, df = 10, which.lam = 10, returnall = FALSE, ...)
}
\arguments{
\item{Y}{response, Nxd matrix}

\item{X}{predictors, Nxp matrix}

\item{df}{the number of spline basis functions to use, defaluts to 10}

\item{which.lam}{component of the tuning parameter vector for saving coefficients}

\item{returnall}{if TRUE, returns extra information, see below}

\item{...}{optional arguments to \code{\link[SAM]{samQL}}}
}
\value{
a list with components
\describe{
\item{knots}{a list with components \code{interior}, a (df-1)xp matrix and \code{boundary}, a 2xd matrix}
\item{B.hat}{a (d)x(1+df*p) matrix containing the retained coefficients (intercept in the first column)}
\item{range.info}{a list with components \code{min} and \code{range} of size px1 each describing the predictors}
\item{spam.out}{if \code{returnall} is TRUE, a list of length d containing all output from \code{\link[SAM]{samQL}}. Useful for the \code{\link{predict}} or \code{\link{fitted}} methods}
\item{X}{the original \code{X} input}
}
}
\description{
Estimates sparse (mulitivariate) additive model
}
\examples{
someData = generateSPASM(100,3,4)
SPAMfit = spam(someData$y, someData$x)
}
