% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spasm.R
\docType{package}
\name{spasm}
\alias{spasm}
\alias{spasm-package}
\title{spasm: A package for implementing sparse additive state space models via the Laplace Gaussian Filter}
\description{
The \code{spasm} package provides three main functions:
\code{\link{lgf1}} which implements the first order filter, \code{\link{spam}} which estimates a sparse additive model with vector-valued response (also with predict method), and \code{\link{generateSPASM}} which generates some data from a state space model with a linear transition equation but non-linear observation equation.
}
