% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_simulations.R
\name{step_C_sim2}
\alias{step_C_sim2}
\title{One realization of the sim2 algorithm from EcoSimR}
\usage{
step_C_sim2(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\description{
https://cran.r-project.org/web/packages/EcoSimR/vignettes/CoOccurrenceVignette.html
}
