% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_simulations.R
\name{step_CA_rowrandom}
\alias{step_CA_rowrandom}
\title{One realization of the an algorithm that resamples each row randomly}
\usage{
step_CA_rowrandom(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\description{
For each row, take the total number of individuals, and
re-assignn each of them with equal probability to each column.
This should actually be equivalent to the Poisson algorithm (i.e. the
distribution of abundances of each species should be Poisson).
}
