% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{V_ratio}
\alias{V_ratio}
\title{Variance ratio of Schluter}
\usage{
V_ratio(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can 
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A single number, the variance ratio.
}
\description{
Variance ratio of Schluter
}
\references{
Schluter, D. 1984. A variance test for detecting species associations, with some example applications. Ecology 65: 998-1005.
}
