% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_measures.r
\name{C_forbes}
\alias{C_forbes}
\title{Forbes' coefficient of association among species in a community matrix}
\usage{
C_forbes(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
Forbes' coefficient of association among species in a community matrix
}
\references{
Forbes S.A. (1907) On the local distribution of certain Illinois
fishes: An essay in statistical ecology. Bulletin of the Illinois State
Laboratory of Natural History, 7: 273-303.

Arita H. (2016) Species co-occurrence analysis: pairwise versus
matrix-level approaches. Global Ecology and Biogeography, 25: 1397-1400.
}
