% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_jacc}
\alias{C_jacc}
\title{Pairwise Jaccard associations among species in a community matrix}
\usage{
C_jacc(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
This is the positive association measure (as opposed to dissimilarity).
Note: Before the calculation, species (rows) with zero occurrences, and sites (columns)
with zero species are removed from m.
}
