% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_pears}
\alias{C_pears}
\title{Pearson tetrachoric correaltions for binary data}
\usage{
C_pears(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
This is the index A_30 from Hubalek (1982), who considered it to be one of
the recommended indices.
Note: Before the calculation, species (rows) with zero occurrences are removed from m.
}
\references{
Hubalek Z. (1982) Coefficients of association and similarity, based
on binary (presence-absence) data: an evaluation. Biol. Rev. 57: 669-689.
}
