% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_seg}
\alias{C_seg}
\title{The classical 'C-score' seggregation metric of Stone & Roberts (1990),
and its scaled version}
\usage{
C_seg(m, lst = FALSE, fun = FALSE, scale = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}

\item{scale}{Should the raw metric be scaled using the total number of possible
site combinations?}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
The classical 'C-score' seggregation metric of Stone & Roberts (1990),
and its scaled version
}
\references{
Stone L. & Roberts A. (1990) The checkerboard score and
species distributions. Oecologia 85: 74-79.

McNickle G.G. et al. (2018) Checkerboard score-area relationships
reveal spatial scales of plant community structure. Oikos 127: 415-426.
}
