% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_incidence_matrices.R
\name{C_w}
\alias{C_w}
\title{Whittaker's index of overall association in a community matrix}
\usage{
C_w(m)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}
}
\value{
A single number, which is the ratio of mean occupancy and the number of
}
\description{
This metric is equivalent to calculating the classical Whittaker's beta diversity on
a transposed matrix. The idea to use this metric for species associations arose
during a discussion with Aniko Toth at IBS 2019 conference in Malaga, and hence
the name of the metric.
}
