% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_abundance_matrices.R
\name{CA_bray}
\alias{CA_bray}
\title{Bray-Curtis dissimilarity matrix}
\usage{
CA_bray(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}
}
\value{
A dist or data.frame objects with the pairwise values.
}
\description{
Calculates Bray-Curtis dissimilarity matrix for a community matrix.
This is a wrapper to a specific call to the 'vegdist' function in 'vegan'.
}
\references{
Legendre P. & De Caceres M. (2013) Beta diversity as the variance of community
data: dissimilarity coefficients and partitioning. Ecology Letters 16: 951-963.
}
