% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_matrix_measures.R
\name{CA_chi}
\alias{CA_chi}
\title{Chi-square distance matrix}
\usage{
CA_chi(m, lst = FALSE, fun = FALSE)
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}
}
\value{
A dist or data.frame objects with the pairwise values.
}
\description{
Calculates chi-square distance matrix for mixed community matrices. In particular,
the abundance data are chi-square-transformed, and then used for calcluation
of Euclidean distances.
}
\references{
Legendre P. & De Caceres M. (2013) Beta diversity as the variance of community
data: dissimilarity coefficients and partitioning. Ecology Letters 16: 951-963.

Legendre P. & Legendre L. (2012) Numerical Ecology. Elsevier. pp 265-335.
}
