% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_for_abundance_matrices.R
\name{CA_cov_cor}
\alias{CA_cov_cor}
\title{Covariance or correlation matrix calculated from community matrix}
\usage{
CA_cov_cor(m, lst = FALSE, fun = FALSE, transf = "hellinger",
  correlation = TRUE, method = "pearson")
}
\arguments{
\item{m}{Community data matrix with species as rows and sites as columns. Can
contain either incidences (1/0) or abundances (natural numbers).}

\item{lst}{Should the results be returned as a 'dist' object (FALSE), or
as a 'data.frame' (TRUE)?}

\item{fun}{Additional function (e.g. log-transformation) to be applied
to all pairwise values.}

\item{transf}{A transformation function to be applied to the community matrix.
The default is Hellinger transformation. Other options from 'vegan' function
'decostand' are applicable.}

\item{correlation}{Should correlations be returned (TRUE) or covariances (FALSE).}

\item{method}{One of 'peason' (default), 'kendall', or 'spearman'.}
}
\value{
A dist or data.frame objects with the pairwise association values.
}
\description{
Calculates covariance or correlation matrix from a community abundance or
incidence matrix where species are rows and sites are columns.
}
