% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{create_fleet}
\usage{
create_fleet(eq_f = NA, length_50_sel = 1, length_95_sel = 2,
  mpa_reaction = "concentrate", price = 1, cost = 0.1, beta = 1.3,
  theta = 0.1, q = 0.001, fleet_model = "constant-effort",
  effort_allocation = "gravity", initial_effort = 100, target_catch = 0,
  fish)
}
\arguments{
\item{initial_effort}{}
}
\value{
a fleet object
}
\description{
create_fleet
}
\examples{
create_fleet(eq_f = 2,length_50_sel = 25, length_95_sel = 27, fish = bluefish)

}
