% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_patterns_simulated_communities.R
\name{PDFtexp}
\alias{PDFtexp}
\title{Truncated exponential probability density function}
\usage{
PDFtexp(d, alpha, dlim = c(0, 1))
}
\arguments{
\item{d}{a vector or scalar representing spatial distance}

\item{alpha}{the main parameter of the function}

\item{dlim}{the two truncation points of the function (default is 0 and 1)}
}
\value{
Probabiliy densities for each of the d values.
}
\description{
This is the probability density function from Keil (2014) PeerJ: https://peerj.com/preprints/446/.
}
