% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_point_patterns.R
\name{plot.PM}
\alias{plot.PM}
\title{Plot the PM classification scheme}
\usage{
plot.PM(
  PM,
  xlim = NULL,
  ylim = NULL,
  labs = FALSE,
  lab.shift = NULL,
  main = NULL
)
}
\arguments{
\item{PM}{A list object returned by the PM.multispec function.}
}
\value{
A ggplot figure.
}
\description{
This function plots the output of the PM.multispec function. Still under
development - use it cautiously.
}
