% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_communities.R
\name{sim.comm.jittered}
\alias{sim.comm.jittered}
\title{Function that simulates a community of several species.}
\usage{
sim.comm.jittered(abund.vect, var.consp, var.intersp)
}
\arguments{
\item{abund.vect}{vector of (integer) abundances of species in a community}

\item{var.consp}{conspecific variance, i.e. spread of points in the master species}

\item{var.intersp}{interspecific variance, i.e. width of the density kernel}
}
\description{
The simulation proceeds as:
1. Generate point pattern for the most abundant (master) species,
   with a given MV variance 'var.consp'
2. Generate smoothed density of the master points, with a
   given gaussian kernel width 'var.intersp'
3. For each other species in the vector, draw points from the smoothed
   density surface and a given abundance
}
